Conflicts:
	gameProgram/Debug/gameProgram.exe
	gameProgram/Debug/gameProgram.ilk
	gameProgram/Debug/gameProgram.pdb
	gameProgram/gameProgram.sdf
	gameProgram/gameProgram.suo
	gameProgram/gameProgram/Debug/CL.read.1.tlog
	gameProgram/gameProgram/Debug/CL.write.1.tlog
	gameProgram/gameProgram/Debug/camera.obj
	gameProgram/gameProgram/Debug/cl.command.1.tlog
	gameProgram/gameProgram/Debug/commandmanager.obj
	gameProgram/gameProgram/Debug/commandteam.obj
	gameProgram/gameProgram/Debug/debugproc.obj
	gameProgram/gameProgram/Debug/demo.obj
	gameProgram/gameProgram/Debug/demoImport.obj
	gameProgram/gameProgram/Debug/drawList.obj
	gameProgram/gameProgram/Debug/drawListManager.obj
	gameProgram/gameProgram/Debug/effectManager.obj
	gameProgram/gameProgram/Debug/fbxModel.obj
	gameProgram/gameProgram/Debug/game.obj
	gameProgram/gameProgram/Debug/gameImport.obj
	gameProgram/gameProgram/Debug/gameProgram.log
	gameProgram/gameProgram/Debug/import.obj
	gameProgram/gameProgram/Debug/instancingBillboard.obj
	gameProgram/gameProgram/Debug/keyboard.obj
	gameProgram/gameProgram/Debug/link.1508-cvtres.read.1.tlog
	gameProgram/gameProgram/Debug/link.1508-cvtres.write.1.tlog
	gameProgram/gameProgram/Debug/link.1508.read.1.tlog
	gameProgram/gameProgram/Debug/link.1508.write.1.tlog
	gameProgram/gameProgram/Debug/list.obj
	gameProgram/gameProgram/Debug/main.obj
	gameProgram/gameProgram/Debug/manager.obj
	gameProgram/gameProgram/Debug/objectBase.obj
	gameProgram/gameProgram/Debug/objectList.obj
	gameProgram/gameProgram/Debug/padX.obj
	gameProgram/gameProgram/Debug/padXManager.obj
	gameProgram/gameProgram/Debug/phase.obj
	gameProgram/gameProgram/Debug/polygon2D.obj
	gameProgram/gameProgram/Debug/renderer.obj
	gameProgram/gameProgram/Debug/result.obj
	gameProgram/gameProgram/Debug/resultImport.obj
	gameProgram/gameProgram/Debug/shader.obj
	gameProgram/gameProgram/Debug/title.obj
	gameProgram/gameProgram/Debug/titleImport.obj
	gameProgram/gameProgram/Debug/updateList.obj
	gameProgram/gameProgram/Debug/vc100.idb
	gameProgram/gameProgram/Debug/vc100.pdb
	gameProgram/gameProgram/Debug/window.obj
	gameProgram/gameProgram/phase/title/title.cpp
	gameProgram/ipch/gameprogram-2f38d934/gameprogram-58c9e9be.ipch

+++ gameProgram/gameProgram/manager/effectManager/effectManager.h
@@ -62,7 +62,7 @@ public:
 		ANIMATION anim;//アニメーションあるならやるよ
 		int animFrame;
 
-		float compErase;
+		int compErase;
 		float animErase;
 
 		float lifeSpan;//今のところ補完フレームと寿命は同じ
